fix(memory-v2): unbreak Test job — extract sparse tokenizer, register rebuild-corpus-stats route, mock BM25 in sim tests
The Test job on main was failing on 8 tests across three independent
breakages introduced by the BM25 rollout (#29344):

1. Five tests that mock embedding-backend.js wholesale via mock.module()
   broke because sparse-bm25.ts now imports SPARSE_VOCAB_SIZE / tokenize /
   tokenHash from embedding-backend, and those exports were missing from
   the mocks. Extract the three primitives into a new sparse-tokenize.ts
   shared by both the legacy TF-only encoder and the BM25 encoder, so
   transitive importers no longer pay attention to embedding-backend
   mocks. No behavior change — same logic, same exported names.

2. guard-tests "every endpoint dispatched in http-server.ts has a policy
   entry in route-policy.ts" failed because memory/v2/rebuild-corpus-stats
   (added in #29344) had no policy. Register it under settings.write
   alongside the other v2 mutating endpoints (backfill, reembed-skills).

3. sim.test "embeds the query text exactly once via dense + sparse
   backends" tracked generateSparseEmbedding calls, but sim.ts now uses
   generateBm25QueryEmbedding from sparse-bm25.ts. Wrap that function in
   the mock instead, capturing the real reference into a local *before*
   registering the mock — ESM live bindings would otherwise route the
   delegated call back into the mocked version and recurse.
